Build Tools: update npm and composer dependencies.
This is mainly to get around a problem with grunt-composer.  That grunt plugin is written in CoffeeScript.  grunt, itself, used to have CoffeeScript as a dependency, but that was removed in grunt 1.2.0.  So, until grunt-composer is updated to add coffeescript as a dependency, we have to do so as a direct devDependency of this plugin.